Update submodule biosoup and unit tests

+++ test/fasta_parser_test.cpp
@@ -7,7 +7,7 @@
 #include "biosoup/sequence.hpp"
 #include "gtest/gtest.h"
 
-std::atomic<std::uint64_t> biosoup::Sequence::num_objects{0};
+std::atomic<std::uint32_t> biosoup::Sequence::num_objects{0};
 
 namespace bioparser {
 namespace test {
